refactor: rename rtscan workflows and update related documentation
