+++ csl/config.py
@@ -10,7 +10,7 @@ ROOT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 # Dictionary with pairs of data source and chromatographic method
 # If data sources are added or modified, the respective mappings need to be changed here and in:
-# + Mapping for RT models in rtscan_functions.check_rtscan.py
+# + Mapping for RT models in rtscan_functions.update_rtscan.py
 DEFAULT_PAIRS_INST_CHROM = {
     'bfg': 'bfg_nts_rp1',
     'uba': 'uba_nts_rp1',
