+++ csl/main_functions/rtscan.py
@@ -2,8 +2,8 @@
 Main function to operate on retention time data saved in the CSL.
 
 Rtscan Classes:
-    check   : Workflow for checking mistakes in retention time (RT) entries and adding missing RT in the CSL.
-    update  : Workflow for recalculating all retention time entries in the CSL.
+    update : Workflow for adding missing non-experimental retention times (RT) and correcting errors in associated entries in the CSL.
+    recalc : Workflow for recalculating and replacing all non-experimental RTs in the CSL.
 """
 
 from csl.utils import validate_file_path, setup_logger
@@ -22,8 +22,8 @@ setup_logger(log_fpath)  # Sets basic logger configuration and adds stream handl
 
 # Define workflow dictionary
 WORKFLOWS = {
-    "check": CheckRtscan,
-    "update": UpdateRtscan
+    "update": UpdateRtscan,
+    "recalc": RecalcRtscan
 }
 
 
@@ -33,8 +33,8 @@ def scan_rt(operation: str, path_csl: str):
 
     Args:
         operation (str) : Type of operation. Choose from:
-            - 'check'   : Checks for mistakes in RT entries and adds missing RT values in the CSL.
-            - 'update'  : Recalculates all RT entries in the CSL.
+            - 'update'  : Adds missing non-experimental retention times using available experimental data and corrects errors in associated entries in the CSL.
+            - 'recalc'  : Recalculates and replaces all non-experimental retention times using available experimental data in the CSL.
         path_csl (str)  : Path to the CSL file.
     """
