+++ csl/rtscan_functions/__init__.py
@@ -1,10 +1,10 @@
 from .operation_rtscan import OperationRtscan
-from .check_rtscan import CheckRtscan
 from .update_rtscan import UpdateRtscan
+from .recalc_rtscan import RecalcRtscan
 
 
 __all__ = [
     "OperationRtscan",
-    "CheckRtscan",
-    "UpdateRtscan"
+    "UpdateRtscan",
+    "RecalcRtscan"
 ]
